requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
streamlit==1.29.0

# Database
pymongo==4.6.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3

# PyTorch (CPU version - install separately first)
# pip install torch==2.1.2 torchvision==0.16.2 --index-url https://download.pytorch.org/whl/cpu

# Hugging Face Ecosystem
huggingface-hub==0.20.2
tokenizers==0.15.0
safetensors==0.4.1
transformers==4.36.2

# LangChain Stack
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20

# Vector Store
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Image Processing
pillow==10.2.0
opencv-python==4.9.0.80

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Data Processing
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0

# Visualization
plotly==5.18.0

# HTTP
requests==2.31.0

# Server
gunicorn==21.2.0
python-multipart==0.0.6
